# publish.go
package mercure

import (
	"context"
	"errors"
	"fmt"
	"io"
	"log/slog"
	"net/http"
	"strconv"
	"strings"

	"go.opentelemetry.io/otel/trace"
)

type updateContextKeyType struct{}

var UpdateContextKey updateContextKeyType //nolint:gochecknoglobals

// reservedTopicSubstring marks topics generated by the hub itself
// (e.g., subscription events). Publishers must not be able to forge those.
const reservedTopicSubstring = "/.well-known/mercure"

// Element-count caps prevent DoS amplification when callers fit many
// topics/matchers inside a request whose byte size is within transport
// limits (Caddy request_body, Go MaxHeaderBytes). maxPublishTopics is held
// tighter than the subscribe-side caps: `mercure.publish: ["*"]` removes
// per-topic authorization, and every topic lands in slog attrs, OTel span
// attrs, and the codec/Redis payload — 64 covers a canonical IRI plus
// alternates.
const (
	maxClaimMatchers = 1000 // mercure.subscribe / mercure.publish claim array
	maxQueryTopics   = 1000 // "topic" query params on subscribe
	maxPublishTopics = 64   // "topic" form fields on publish
)

// Byte-length caps on publisher-supplied SSE fields. Bytes, not codepoints —
// the SSE wire is byte-oriented and downstream cost (Redis payload, slog
// attrs, OTel span attrs) is byte-billed. 1024 leaves generous headroom for
// IRI-style ids and event-type names (typical IRIs stay under 256 bytes;
// UUIDv7 URNs are 47). Body size is the operator's responsibility via Caddy's
// `request_body` directive.
const (
	maxUpdateIDBytes    = 1024
	maxUpdateTypeBytes  = 1024
	maxUpdateTopicBytes = 1024
)

// Sentinel errors returned by Validate (surfaced through Publish). Callers can
// branch on them via errors.Is; PublishHandler maps them to HTTP status codes.
var (
	ErrReservedTopic    = errors.New(`topic value references the reserved "/.well-known/mercure" namespace`)
	ErrTooManyTopics    = errors.New("too many topics in update")
	ErrTopicTooLong     = errors.New("topic value exceeds the maximum length")
	ErrInvalidTopic     = errors.New("topic value contains a forbidden control character")
	ErrInvalidEventID   = errors.New(`"id" field contains a forbidden control character`)
	ErrInvalidEventType = errors.New(`"type" field contains a forbidden control character`)
	ErrEventIDTooLong   = errors.New(`"id" field exceeds the maximum length`)
	ErrEventTypeTooLong = errors.New(`"type" field exceeds the maximum length`)
)

// ErrPublishTimeout is the cause attached to the dispatch context when a
// configured publish_timeout expires. PublishHandler matches it via
// context.Cause so that only a publish_timeout abort maps to 504 — a
// context.DeadlineExceeded arriving from any other source (a parent context, a
// transport-internal deadline) keeps its normal mapping.
var ErrPublishTimeout = errors.New("publish dispatch exceeded publish_timeout")

// sseFieldForbiddenChars, if copied into SSE id/event/topic fields, would let
// a publisher inject arbitrary SSE fields into subscribers' streams. Event.String
// writes id/type directly into "id: …\n" / "event: …\n" without escaping (only
// "data" is escaped), so CR/LF would forge SSE frames. NUL is rejected as
// defense-in-depth — no legitimate use here, and downstream C-string-style
// parsers truncate on it.
const sseFieldForbiddenChars = "\x00\r\n"

// validateTopics enforces the topic-count and per-topic rules. PublishHandler
// runs it before canDispatch so oversize/forbidden topics are rejected ahead of
// the O(topics × selectors) authorization match; Validate runs it again for the
// programmatic Hub.Publish path.
func validateTopics(topics []string) error {
	if len(topics) > maxPublishTopics {
		return ErrTooManyTopics
	}

	for _, t := range topics {
		switch {
		case len(t) > maxUpdateTopicBytes:
			return ErrTopicTooLong
		case strings.ContainsAny(t, sseFieldForbiddenChars):
			return ErrInvalidTopic
		case strings.Contains(t, reservedTopicSubstring):
			return fmt.Errorf("%q: %w", t, ErrReservedTopic)
		}
	}

	return nil
}

// Validate is the single source of truth for publish-side input rules, now
// exported (upstream #1266). Called from Hub.Publish, it covers every dispatch
// path: the HTTP PublishHandler, the programmatic API, and transports that publish
// through Hub.Publish. A caller that builds an Update from untrusted input and
// dispatches it directly, bypassing Hub.Publish, MUST call Validate first: it
// rejects topic count/length, CR/LF/NUL injection into id/type (CWE-93), and the
// reserved "/.well-known/mercure" namespace. That last rejection is why it is for
// publisher input, not hub-internal subscription events — see ValidateSSEFields
// for the out-of-process transport path.
func (u *Update) Validate() error {
	if err := validateTopics(u.Topics); err != nil {
		return err
	}

	if len(u.ID) > maxUpdateIDBytes {
		return ErrEventIDTooLong
	}

	if !ValidSSEFieldValue(u.ID) {
		return ErrInvalidEventID
	}

	if len(u.Type) > maxUpdateTypeBytes {
		return ErrEventTypeTooLong
	}

	if !ValidSSEFieldValue(u.Type) {
		return ErrInvalidEventType
	}

	return nil
}

// ValidateSSEFields checks the two fields Event.String writes verbatim into the
// server-sent-event wire frame — the "id:" line and the "event:" (Type) line —
// for a CR, LF, or NUL. Those bytes forge frame boundaries and let forged
// content inject arbitrary SSE fields into a subscriber's stream (CWE-93). Data
// is line-escaped by Event.String (event.go dataReplacer) and Topics never reach
// the wire, so neither is checked here — keep this in sync with event.go if the
// frame ever grows a new verbatim field.
//
// The publish path is already covered: Hub.Publish calls Validate, a superset.
// This narrow, exported check exists for a Transport that reads updates from an
// out-of-process backend and must guard the SSE-injection vector before dispatch
// WITHOUT rejecting hub-internal subscription events, which ride reserved topics
// that the full Validate rejects by design. See redistransport decodeStreamEntry.
func (u *Update) ValidateSSEFields() error {
	if !ValidSSEFieldValue(u.ID) {
		return ErrInvalidEventID
	}

	if !ValidSSEFieldValue(u.Type) {
		return ErrInvalidEventType
	}

	return nil
}

// ValidSSEFieldValue reports whether s is safe to write verbatim as a server-sent
// event field value — i.e. it holds no CR, LF, or NUL that would forge SSE frame
// boundaries (CWE-93). ValidateSSEFields applies it to an update's id and type; a
// Transport can also apply it to SSE-bound metadata it sources from an untrusted
// backend, such as a history-replay cursor destined for the Last-Event-Id
// response header, which is not covered by validating the decoded update alone.
func ValidSSEFieldValue(s string) bool {
	return !strings.ContainsAny(s, sseFieldForbiddenChars)
}

// writePublishError maps a Validate/Publish sentinel error to its HTTP status:
// 403 for reserved-topic forgery, 400 for the length/count/character failures,
// 500 otherwise. The publish_timeout 504 is handled in PublishHandler (it keys
// on the dispatch context's cause, which writePublishError cannot see).
func writePublishError(w http.ResponseWriter, err error) {
	switch {
	case errors.Is(err, ErrReservedTopic):
		http.Error(w, err.Error(), http.StatusForbidden)
	case errors.Is(err, ErrTooManyTopics),
		errors.Is(err, ErrTopicTooLong),
		errors.Is(err, ErrInvalidTopic),
		errors.Is(err, ErrInvalidEventID),
		errors.Is(err, ErrInvalidEventType),
		errors.Is(err, ErrEventIDTooLong),
		errors.Is(err, ErrEventTypeTooLong):
		http.Error(w, err.Error(), http.StatusBadRequest)
	default:
		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
	}
}

// Publish broadcasts the given update to all subscribers.
// The id field of the Update instance can be updated by the underlying Transport.
//
// Both the HTTP PublishHandler and programmatic callers converge here, and
// Validate runs first (see its doc for what it rejects), so every dispatch path —
// including transports that publish through this method — is guarded before an
// update reaches a subscriber.
func (h *Hub) Publish(ctx context.Context, update *Update) error {
	ctx, span := startSpan(ctx, "mercure.publish", trace.WithSpanKind(trace.SpanKindProducer))

	var validated bool
	// Deferred so the ID assigned by the transport via AssignUUID lands on the
	// span. Gated on validation: a rejected update may carry oversize or
	// CR/LF/NUL id/topics that the byte caps exist to keep out of traces.
	defer func() {
		if validated && span.IsRecording() {
			span.SetAttributes(update.SpanAttributes()...)
		}

		span.End()
	}()

	if err := update.Validate(); err != nil {
		if h.logger.Enabled(ctx, slog.LevelInfo) {
			h.logger.LogAttrs(ctx, slog.LevelInfo, "Rejected invalid update", slog.Any("error", err))
		}

		recordSpanError(span, err)
		h.recordPublishFailure(update, PublishFailureReasonValidation)

		return err
	}

	validated = true

	ctx = context.WithValue(ctx, UpdateContextKey, update)

	if err := h.transport.Dispatch(ctx, update); err != nil {
		if h.logger.Enabled(ctx, slog.LevelError) {
			// mercureHandler injects bounded update attrs via UpdateContextKey
			// + Update.LogValue (id/retry/topics/private/data_bytes; type and
			// data behind Debug). No explicit update fields needed here.
			h.logger.LogAttrs(ctx, slog.LevelError, "Failed to dispatch update", slog.Any("error", err))
		}

		recordSpanError(span, err)
		h.recordPublishFailure(update, publishFailureReasonForDispatch(ctx, err))

		return err //nolint:wrapcheck
	}

	h.metrics.UpdatePublished(update)

	if h.logger.Enabled(ctx, slog.LevelDebug) {
		h.logger.LogAttrs(ctx, slog.LevelDebug, "Update published")
	}

	return nil
}

// recordPublishFailure reports a failed publish to the optional
// PublishFailureReporter extension of h.metrics, when implemented. The
// type-assertion is cheap and stays off the success path (UpdatePublished),
// so a Metrics impl that doesn't care about failures is unaffected.
func (h *Hub) recordPublishFailure(u *Update, reason PublishFailureReason) {
	if r, ok := h.metrics.(PublishFailureReporter); ok {
		r.UpdatePublishFailed(u, reason)
	}
}

// isPublishTimeout reports whether a Hub.Publish dispatch error is a
// publish_timeout abort: the dispatch context's cause is ErrPublishTimeout AND
// the returned error is itself a deadline error (context.DeadlineExceeded) or
// the ErrPublishTimeout sentinel. Anchoring on the cause — not a bare
// DeadlineExceeded — keeps an unrelated context's deadline from being mistaken
// for a publish_timeout. PublishHandler (→ 504) and publishFailureReasonForDispatch
// (→ reason=timeout) share this single predicate so the two cannot drift apart
// and the timeout metric tracks 504s exactly.
func isPublishTimeout(ctx context.Context, err error) bool {
	return errors.Is(context.Cause(ctx), ErrPublishTimeout) &&
		(errors.Is(err, context.DeadlineExceeded) || errors.Is(err, ErrPublishTimeout))
}

// publishFailureReasonForDispatch classifies a transport.Dispatch error for the
// failure metric: a publish_timeout abort (see isPublishTimeout) is timeout;
// every other dispatch failure — transport unreachable, XADD error, codec error,
// and any deadline error NOT tagged with the ErrPublishTimeout cause — falls into
// the transport catch-all.
func publishFailureReasonForDispatch(ctx context.Context, err error) PublishFailureReason {
	if isPublishTimeout(ctx, err) {
		return PublishFailureReasonTimeout
	}

	return PublishFailureReasonTransport
}

// PublishHandler allows publisher to broadcast updates to all subscribers.
//
//nolint:funlen,gocognit,gocyclo
func (h *Hub) PublishHandler(w http.ResponseWriter, r *http.Request) {
	ctx, span := startSpan(r.Context(), "mercure.publish", trace.WithSpanKind(trace.SpanKindProducer))

	var (
		u         *Update
		published bool
	)
	// Deferred so the ID assigned by the transport via AssignUUID lands on the
	// span. Gated on a successful publish so a rejected update's unvalidated
	// id/topics never reach the tracing exporter.
	defer func() {
		if published && span.IsRecording() {
			span.SetAttributes(u.SpanAttributes()...)
		}

		span.End()
	}()

	r = r.WithContext(ctx)

	var claims *claims

	if h.publisherJWTKeyFunc != nil {
		var err error

		claims, err = h.authorizeAndBind(r, true)
		if err != nil || claims == nil || claims.Mercure.Publish == nil {
			h.httpAuthorizationError(w, r, err)

			if err != nil {
				recordSpanError(span, err)
			}

			return
		}
	}

	if r.ParseForm() != nil { //nolint:gosec // body size can be limited using Caddy's request_body directive
		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)

		return
	}

	topics := r.PostForm["topic"]
	if len(topics) == 0 {
		http.Error(w, `Missing "topic" parameter`, http.StatusBadRequest)

		return
	}

	// Run before canDispatch (the publisher is already authenticated) so
	// oversize/forbidden/reserved topics are rejected ahead of the selector
	// match; Validate re-runs validateTopics on the programmatic path. This early
	// reject never reaches Hub.Publish, so meter the validation failure here too.
	if err := validateTopics(topics); err != nil {
		writePublishError(w, err)
		recordSpanError(span, err)
		h.recordPublishFailure(&Update{Topics: topics}, PublishFailureReasonValidation)

		return
	}

	var retry uint64

	if retryString := r.PostForm.Get("retry"); retryString != "" {
		var err error
		if retry, err = strconv.ParseUint(retryString, 10, 64); err != nil {
			http.Error(w, `Invalid "retry" parameter`, http.StatusBadRequest)

			return
		}
	}

	// id/type length and CR/LF/NUL validation runs in Validate (via
	// Hub.Publish), which maps failures to the sentinel errors below.
	rawID := r.PostForm.Get("id")
	rawType := r.PostForm.Get("type")

	private := len(r.PostForm["private"]) != 0
	if claims != nil && !canDispatch(h.topicSelectorStore, topics, claims.Mercure.Publish) { //nolint:nestif
		if private {
			http.Error(w, http.StatusText(http.StatusUnauthorized), http.StatusUnauthorized)

			return
		}

		infoEnabled := h.logger.Enabled(ctx, slog.LevelInfo)
		if h.isBackwardCompatiblyEnabledWith(7) {
			if infoEnabled {
				h.logger.LogAttrs(ctx, slog.LevelInfo, `Deprecated: posting public updates to topics not listed in the "mercure.publish" JWT claim is deprecated since the version 7 of the protocol, use '["*"]' as value to allow publishing on all topics.`)
			}
		} else {
			if infoEnabled {
				h.logger.LogAttrs(ctx, slog.LevelInfo, `Unsupported: posting public updates to topics not listed in the "mercure.publish" JWT claim is not supported anymore, use '["*"]' as value to allow publishing on all topics or enable backward compatibility with the version 7 of the protocol.`)
			}

			http.Error(w, http.StatusText(http.StatusUnauthorized), http.StatusUnauthorized)

			return
		}
	}

	u = &Update{
		Topics:  topics,
		Private: private,
		Debug:   h.debug,
		Event:   Event{r.PostForm.Get("data"), rawID, rawType, retry},
	}

	// Detach from the request context so a publisher disconnecting mid-publish
	// does not abort the dispatch (the update would otherwise be lost). When a
	// publish timeout is configured, bound the now-detached dispatch so a stalled
	// transport cannot block this goroutine indefinitely. WithTimeoutCause tags
	// the expiry with ErrPublishTimeout so the 504 below fires only for this
	// handler's deadline, not a context.DeadlineExceeded from some other source.
	dispatchCtx := context.WithoutCancel(ctx)

	if h.publishTimeout > 0 {
		var cancel context.CancelFunc

		dispatchCtx, cancel = context.WithTimeoutCause(dispatchCtx, h.publishTimeout, ErrPublishTimeout)
		defer cancel()
	}

	// Validation, dispatch, logging and metrics live in Hub.Publish.
	if err := h.Publish(dispatchCtx, u); err != nil {
		// A publish_timeout abort surfaces as the dispatch context's cause. The
		// write may already have committed, so it is indeterminate, not a clean
		// failure: return 504 and tell the publisher a retry is safe only for an
		// idempotent (stable-id) update. Every other error keeps its mapping.
		// Require the returned error to be a deadline error too, not just the
		// context state: a fast non-timeout error must not be reclassified as 504
		// if the deadline happens to expire in the gap before this check.
		if isPublishTimeout(dispatchCtx, err) {
			http.Error(w, "publish timed out; the update may already have been published — retry only with a stable, publisher-supplied id to avoid duplicates", http.StatusGatewayTimeout)
		} else {
			writePublishError(w, err)
		}

		// Mirror the error onto the handler span too; Hub.Publish's child
		// span already records it, but leaving the parent span as success
		// is misleading.
		recordSpanError(span, err)

		return
	}

	published = true

	if _, err := io.WriteString(w, u.ID); err != nil {
		if h.logger.Enabled(ctx, slog.LevelInfo) {
			h.logger.LogAttrs(ctx, slog.LevelInfo, "Failed to write publish response", slog.Any("error", err))
		}

		return
	}
}