# spec/openapi.yaml
openapi: 3.0.3
info:
  title: The Mercure protocol
  description: |
    [Mercure](https://mercure.rocks) is a protocol providing a
    publish-subscribe mechanism for public and private web resources. It
    enables pushing web content to browsers and other clients in a fast,
    reliable, and battery-efficient way. It is especially useful for
    publishing real-time updates of resources served through sites and web
    APIs to web and mobile applications.
  version: 0.3.2
  contact:
    name: Kévin Dunglas
    url: https://mercure.rocks
    email: contact@mercure.rocks
  license:
    name: GNU Affero General Public License v3.0 or later
    url: https://github.com/dunglas/mercure/blob/master/LICENSE
    x-spdx: AGPL-3.0-or-later
servers:
  - url: "{scheme}://{host}"
    variables:
      scheme:
        default: https
        enum:
          - https
      host:
        default: localhost
...
paths:
  "/.well-known/mercure":
    get:
      operationId: subscribe
      tags:
        - Subscribe
      summary: Subscribe to updates
      description: |
        Opens a Server-Sent Events connection to receive updates for the
        specified topics.
      externalDocs:
        description: Subscription specification
        url: https://mercure.rocks/spec#subscription
      parameters:
        - name: topic
          in: query
          description: |
            The topic selectors to subscribe to. Can be any string;
            URI Templates (RFC 6570) and `*` (all topics) are supported.
          required: true
          style: form
          explode: true
          schema:
            type: array
            minItems: 1
            items:
              type: string
          examples:
            singleTopic:
              summary: Subscribe to a single topic
              value: ["https://example.com/books/1"]
            uriTemplate:
              summary: Subscribe using a URI template
              value: ["https://example.com/books/{id}"]
            wildcard:
              summary: Subscribe to all topics
              value: ["*"]
        - name: lastEventID
          in: query
          description: |
            The last received event ID, to retrieve missed events. If both
            this parameter and the `Last-Event-ID` header are present, the
            header MUST take precedence. The reserved value `earliest` hints
            the hub to send all available events; the hub MAY or MAY NOT
            fulfill this request.
          schema:
            type: string
          example: "urn:uuid:5e94c686-2c0b-4f9b-958c-92ccc3bbb4eb"
        - name: Last-Event-ID
          in: header
          description: |
            The last received event ID. MUST take precedence over the
            `lastEventID` query parameter.
          schema:
            type: string
      responses:
        "200":
          description: Event stream opened.
          headers:
            Last-Event-ID:
              description: |
                The ID of the event preceding the first event sent to the
                subscriber, or `earliest` if there is no preceding event.
                Only present when the request includes a `lastEventID`
                parameter or `Last-Event-ID` header.
              schema:
                type: string
          content:
            "text/event-stream": {}
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
      security:
        - Bearer: []
        - Cookie: []
        - QueryParameter: []
        - {}
    post:
      operationId: publish
      tags:
        - Publish
      summary: Publish an update
      description: |
        Publishes an update to the hub, which dispatches it to subscribers
        whose topic selectors match the update's topics.
      externalDocs:
        description: Publishing specification
        url: https://mercure.rocks/spec#publication
      requestBody:
        required: true
        content:
          "application/x-www-form-urlencoded":
            schema:
              type: object
              properties:
                topic:
                  description: |
                    The identifiers of the updated topic. It is RECOMMENDED to
                    use an IRI. If present several times, the first occurrence
                    is considered the canonical IRI and others are considered
                    alternate IRIs. The hub MUST dispatch this update to
                    subscribers subscribed to either canonical or alternate IRIs.
                  type: array
                  minItems: 1
                  items:
                    type: string
                data:
                  description: The content of the new version of this topic.
                  type: string
                private:
                  description: |
                    If this name is set, the update MUST NOT be dispatched to
                    subscribers not authorized to receive it. It is recommended
                    to set the value to `on`, but it can contain any value
                    including an empty string.
                  type: string
                id:
                  description: |
                    The topic's revision identifier, used as the SSE's `id`
                    property. MUST NOT start with `#`. SHOULD be a valid IRI.
                    If omitted, the hub MUST generate a valid IRI. A UUID or
                    DID MAY be used. The hub MAY also generate a relative URI
                    composed of a fragment. The hub MAY ignore the provided ID
                    and generate its own.
                  type: string
                type:
                  description: The SSE's `event` property (a specific event type).
                  type: string
                retry:
                  description: |
                    The SSE's `retry` property (the reconnection time, in
                    milliseconds). Non-negative integer; the hub rejects
                    negative or non-numeric values with a `400` response.
                  type: integer
                  format: int64
                  minimum: 0
              required:
                - topic
            encoding:
              topic:
                style: form
                explode: true
            examples:
              publicUpdate:
                summary: A public update
                value:
                  topic:
                    - "https://example.com/books/1"
                  data: '{"@id": "/books/1", "title": "Updated Title"}'
              privateUpdate:
                summary: A private update
                value:
                  topic:
                    - "https://example.com/books/1"
                  data: '{"@id": "/books/1", "title": "Confidential"}'
                  private: "on"
      responses:
        "200":
          description: The `id` generated by the hub for this update.
          content:
            "text/plain":
              schema:
                type: string
              example: "urn:uuid:e1ee88e2-532a-4d6f-ba70-f0f8bd584022"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
      security:
        - Bearer: []
        - Cookie: []
        - QueryParameter: []
...
components:
...
  responses:
    "400":
      description: Missing or invalid required parameters.
    "401":
      description: Not authorized (missing or invalid JWS, or insufficient permissions for the requested topic).
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        SHOULD be used by non-browser clients. When present, the hub MUST
        use this credential and ignore the `authorization` query parameter
        and the `mercureAuthorization` cookie.
    Cookie:
      type: apiKey
      in: cookie
      name: mercureAuthorization
      description: |
        SHOULD be used by web browsers. The cookie is RECOMMENDED to be
        named `mercureAuthorization` but may differ to prevent conflicts
        when using multiple hubs on the same domain. Used only when no
        `Authorization` header and no `authorization` query parameter are
        provided.
    QueryParameter:
      type: apiKey
      in: query
      name: authorization
      description: |
        SHOULD NOT be used unless it is impossible to use the Authorization
        header or a secure cookie. Hubs MAY support this method. When
        present and no `Authorization` header is provided, the hub MUST
        use this credential and ignore the cookie.
security:
  - Bearer: []
  - Cookie: []
  - QueryParameter: []
...